Basic support for the unikernel HermitCore: HermitCore supports C, C++, Fortran and Go; available providers are QEMU, uHyve and OpenStack.
HermitCore relies on a proxy (hermitcore_proxy folder) when using QEMU
or uHyve as the provider. This proxy is started by UniK when an instance
is launched. Then, the proxy handles the start of the HermitCore
instance (e.g., by using QEMU). If the proxy receives a SIGTERM, it
terminates HermitCore's QEMU (or uHyve) instance.

HermitCore's logging (in the case of QEMU or uHyve):
HermitCore is started by a proxy. This proxy logs HermitCore's output
in a file in the /tmp directory. HermitCore's proxy also serves HTTP
requests for retrieving the logs when QEMU or uHyve is used as the
provider.
Therefore, get_instance_logs.go (QEMU) was adapted to send a HTTP
request to HermitCore's proxy. As a consequence, the instance needs an
IP address while running such that the logs can be retrieved. This IP
is initially set in daemon.go in the RunInstanceParams. There, the IP
is set by examining the IP that is used by the client to reach the
daemon.

redir option in QEMU:
For adding the possibility of redirecting a port in QEMU (e.g. in case
of a simple server example) the option 'redir_port' was added to the
daemon-config.yaml. For instance by setting 'redir_port: 8080' in the
config file, the port will be redirected in case of HermitCore's
proxy being used.

Add uHyve to the available providers:
If uHyve is specified as the provider, HermitCore's proxy will not
start HermitCore by using QEMU. Instead, it will use uHyve, a thin
hypervisor powered by Linux's KVM API.

OpenStack support:
In order to enable OpenStack support for HermitCore in UniK, a container
was added which compiles the sources into an image and then build an ISO
image using grub-mkrescue. Checks were added in stage.go such that the
disk format is set to 'iso' instead of 'qcow2' in the case of
HermitCore's ISO image.

Documentation:
README.md: Documentation of HermitCore added

./compilers/hermitcore.md: Added examples of using HermitCore with C,
C++, Go and Fortran.

./providers/uhyve.md: Added short description of uHyve